refactor(core): fold skill-bridge + collection-watchers into @mulmoclaude/core
Eliminates the last two standalone packages/services/* packages by folding
them into @mulmoclaude/core as subpath exports, so they no longer need to be
published separately — they ship inside core.

- packages/services/skill-bridge → @mulmoclaude/core/skill-bridge
- packages/services/collection-watchers → @mulmoclaude/core/collection-watchers
- Their @mulmoclaude/core/* imports become intra-package relative imports.
  Cross-subsystem singletons (the notifier instance, the collection host) stay
  shared because vite extracts ONE shared chunk across entries — verified: both
  notifier/index.js and collection-watchers/index.js import configureNotifier
  from the same dist chunk; the test_reconciler unit test (configures notifier,
  asserts the watcher's publishes land) passes.
- core 0.1.0 → 0.2.0 (new public subpaths); packages/mulmoclaude drops the two
  deps and requires @mulmoclaude/core ^0.2.0.
- packages/services/ is now empty → removed its build-workspaces step from
  build:packages / build:packages:dev.

packages/services/ now holds nothing; core absorbed all 7 original shared
services. Only @mulmoclaude/collection-plugin (Vue) remains a separate package.

Verified green: build, typecheck, lint, format, unit tests (core 38/38),
e2e (439 passed).

NOTE: requires publishing @mulmoclaude/core@0.2.0 (with the new subpaths)
before this can install from the registry — smoke will 404 until then.
