+++ packages/mulmoclaude/package.json
@@ -35,12 +35,10 @@
     "@mulmochat-plugin/ui-image": "^0.4.0",
     "@mulmoclaude/chart-plugin": "^0.1.1",
     "@mulmoclaude/collection-plugin": "^0.5.1",
-    "@mulmoclaude/collection-watchers": "^0.1.0",
-    "@mulmoclaude/core": "^0.1.0",
+    "@mulmoclaude/core": "^0.2.0",
     "@mulmoclaude/form-plugin": "^0.1.0",
     "@mulmoclaude/html-plugin": "^0.2.0",
     "@mulmoclaude/markdown-plugin": "^0.1.2",
-    "@mulmoclaude/skill-bridge": "^0.1.0",
     "@mulmoclaude/spotify-plugin": "^0.1.0",
     "@mulmoclaude/x-plugin": "^0.1.1",
     "@receptron/task-scheduler": "^0.1.0",
