--- packages/services/collection-watchers/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@mulmoclaude/collection-watchers",
-  "version": "0.1.3",
-  "description": "Collection-completion bell subsystem (fs.watch plumbing + convergent reconciler) shared by MulmoClaude and MulmoTerminal. Depends on the @mulmoclaude/core/notifier singleton + @mulmoclaude/core/collection; the host injects its notification taxonomy + in-app routing via a CollectionNotificationAdapter.",
-  "type": "module",
-  "main": "./dist/index.js",
-  "module": "./dist/index.js",
-  "types": "./dist/index.d.ts",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs",
-      "default": "./dist/index.js"
-    }
-  },
-  "files": [
-    "dist"
-  ],
-  "publishConfig": {
-    "access": "public"
-  },
-  "scripts": {
-    "build": "vite build",
-    "typecheck": "tsc --noEmit",
-    "test": "tsx --test test/test_*.ts"
-  },
-  "peerDependencies": {
-    "@mulmoclaude/core": "*"
-  },
-  "devDependencies": {
-    "@mulmoclaude/core": "*",
-    "@types/node": "^26.0.1",
-    "tsx": "^4.22.4",
-    "typescript": "^6.0.3",
-    "vite": "^8.0.13",
-    "vite-plugin-dts": "^5.0.0"
-  },
-  "license": "MIT"
-}
